aibot_bug/requirements.txt
google-generativeai
requests
tabulate
colorama
beautifulsoup4
urllib3
aiohttp
python-dotenv